test(gateway): align create coverage with harness
